+++ verdeciel/Porthole.Above.swift
@@ -18,7 +18,7 @@ class PanelAbove : Panel
 		while i < 3
 		{
 			let test = SCNLine(vertices: [SCNVector3(0,templates.radius * -2,0.75), SCNVector3(0,templates.radius * -2,0.95)], color: white)
-			test.eulerAngles.y = (degToRad(CGFloat(120 * i)))
+			test.eulerAngles.y = (degToRad(Float(120 * i)))
 			aim.addChildNode(test)
 			aim.eulerAngles.y = degToRad(240)
 			i += 1
